// movil/src/app/models/pedidos/informes.tecnicos.response.ts
// Generated by https://quicktype.io

export interface InformesTecnicosResponse {
    code:     number;
    status:   string;
    informes: InformeTecnico[];
}

export interface InformeTecnico {
    id:                     number;
    observaciones:          string;
    fecha_salida:           string;
    nombre:                 string;
    municipio:              string;
    departamento:           string;
    informeTecnico:         string;
    archivo_pcr:            string;
    histopatologia:         string;
    laboratorioNutricional: string;
}